1. fix the uint64_t bugs

+++ include/CPPAlgos/BCRSCPPAlgo.h
@@ -33,7 +33,7 @@ namespace AMMBench {
          * @param sketchSize the size of sketc or sampling
          * @return the output c matrix
          */
-        virtual torch::Tensor amm(torch::Tensor A, torch::Tensor B, int sketchSize);
+        virtual torch::Tensor amm(torch::Tensor A, torch::Tensor B, uint64_t sketchSize);
 
     };
